+++ test/cases/entry-js-pug/webpack.config.js
@@ -34,7 +34,7 @@ module.exports = {
         test: /\.pug$/,
         loader: PugPlugin.loader,
         options: {
-          method: 'render',
+          //method: 'render',
         },
       },
     ],